Uitwerkingen 1e les React tevens gelijk alles opgesplits om de App. js schoon te houden
