--- src/App.css
@@ -1,80 +1,0 @@
-main {
-    margin: 24px 0;
-    display: flex;
-    justify-content: center;
-}
-
-nav {
-    margin: 24px 0;
-}
-
-footer {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
-section {
-    background-color: #F6F3EF;
-    font-size: 10px;
-    height: 400px;
-    overflow: hidden;
-    flex-basis: 350px;
-    flex-grow: 1;
-    flex-shrink: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-section p {
-    max-width: 200px;
-}
-
-button {
-    border: 1px solid #BC4C2A;
-    color:  #BC4C2A;
-    background-color: transparent;
-    font-size: 10px;
-    padding: 8px 16px;
-    text-transform: lowercase;
-    cursor: pointer;
-}
-
-button:hover {
-    background-color: #BC4C2A;
-    color: white;
-    transition: all .4s ease-in-out;
-}
-
-button:disabled {
-    cursor: not-allowed;
-    border-color: #D08169;
-    color: #D08169;
-}
-
-button + button {
-    margin-left: 12px;
-}
-
-article {
-    width: 100px;
-    position: relative;
-    text-align: center;
-    margin: 16px;
-}
-
-article span {
-    background-color: #BC4C2A;
-    position: absolute;
-    top: 0;
-    left: 0;
-    color: white;
-    font-size: 8px;
-    padding: 4px;
-}
-
-article p {
-    font-size: 10px;
-}
