+++ src/components/Modules/Buttons/Buttons.jsx
@@ -1,0 +1,15 @@
+import "./buttons.css"
+import React from "react";
+
+export default function Buttons({ children, clickHandler, isDisabled }) {
+    return (
+        <button
+            type="button"
+            disabled={isDisabled}
+            onClick={clickHandler}
+        >
+            { children }
+        </button>
+    );
+}
+
