+++ src/components/Modules/Tiles/tiles.css
@@ -1,0 +1,17 @@
+tiles {
+    background-color: #F6F3EF;
+    font-size: 10px;
+    height: 400px;
+    overflow: hidden;
+    flex-basis: 350px;
+    flex-grow: 1;
+    flex-shrink: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+tiles p {
+    max-width: 200px;
+}
