+++ src/components/Pages/Footer-Info/footertyles.css
@@ -1,0 +1,5 @@
+footer-info {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+}
