+++ src/components/Pages/Main-Products/main-products.css
@@ -1,0 +1,5 @@
+main-product {
+    margin: 24px 0;
+    display: flex;
+    justify-content: center;
+}
