+++ src/components/Pages/Nav-Top/nav-top.css
@@ -1,0 +1,3 @@
+nav {
+    margin: 24px 0;
+}
